Start an hsearch branch of whoosher
The purpose of this branch is to explore making whoosher work
effectively in modern tiddlyweb situations, including providing
a way to be an auxilliary search system, alongside whatever the
store being used, uses.

This commit:

* changes the default schema so that there is scoring
* removes extraneous fields
* adds a creator field
* removes 'index_query'[1]
* gets rid of irrelevant query_dict whooha
* adds support for either replacing or adding a router handler that
  provides the search

[1] Various performance improvements revealed by TiddlySpace indicate
that index_query is no longer very useful. It was only ever useful in
large installations, and the benefits of using memcached and other
optimizations in large installations clobber the use of index_query,
so...