Fix a subtle bug with tag indexing.
Under certain conditions a phrase search could trigger an
error when investigating tags. Tags, as keywords, do not do
phrase searches though, so if the parser has set up to do
a "positional" search things break. The fix is to make sure
that a quoted search phrase isn't passed to the tag part of the
index as a phrase, but only as a keyword.

Finding this was hard because the error message isn't clear
and parts of the documentation talk phrase, some positional.
And I couldn't create a failing test. I had to use the failing
>800MB index.

In the end the cause was that a default was misset in the schema
for tags.